requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
requests==2.32.3
streamlit==1.50.0
python-dotenv==1.0.1

# Gemini SDK
google-genai==1.47.0

# USB serial communication
pyserial==3.5

# Data / charts
pandas==2.2.2